requirements.txt
Django==5.2.8
psycopg2-binary==2.9.11
python-dotenv==1.0.0
dj-database-url==2.1.0
gunicorn==23.0.0
whitenoise==6.8.2
djangorestframework==3.15.2
